dashboard/requirements.txt
pandas
streamlit
streamlit-autorefresh
altair
plotly
python-dotenv
psycopg2-binary